Create README.md
I would like to ask those of you that took the time to read my profile, if you would please copy and paste this "The-Restore-Hope-Foundation" into your social page and  post it to share it, I have a goal of taking the Foundation global by Christmas Day.   I'm hoping I can count on y'all to help with this goal.    Thank you very much Hope@Restore-Hope-Foundation